+++ src/Twilio/Rest/Taskrouter/V1/Workspace/TaskQueueResource.cs
@@ -334,6 +334,7 @@ namespace Twilio.Rest.Taskrouter.V1.Workspace
         /// <param name="friendlyName"> The `friendly_name` of the TaskQueue resources to read. </param>
         /// <param name="evaluateWorkerAttributes"> The attributes of the Workers to read. Returns the TaskQueues with Workers that match the attributes specified in this parameter. </param>
         /// <param name="workerSid"> The SID of the Worker with the TaskQueue resources to read. </param>
+        /// <param name="ordering"> Sorting parameter for TaskQueues </param>
         /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <param name="limit"> Record limit </param>
@@ -343,11 +344,12 @@ namespace Twilio.Rest.Taskrouter.V1.Workspace
                                                      string friendlyName = null,
                                                      string evaluateWorkerAttributes = null,
                                                      string workerSid = null,
+                                                     string ordering = null,
                                                      int? pageSize = null,
                                                      long? limit = null,
                                                      ITwilioRestClient client = null)
         {
-            var options = new ReadTaskQueueOptions(pathWorkspaceSid){ FriendlyName = friendlyName, EvaluateWorkerAttributes = evaluateWorkerAttributes, WorkerSid = workerSid, PageSize = pageSize, Limit = limit};
+            var options = new ReadTaskQueueOptions(pathWorkspaceSid){ FriendlyName = friendlyName, EvaluateWorkerAttributes = evaluateWorkerAttributes, WorkerSid = workerSid, Ordering = ordering, PageSize = pageSize, Limit = limit};
             return Read(options, client);
         }
 
@@ -357,6 +359,7 @@ namespace Twilio.Rest.Taskrouter.V1.Workspace
         /// <param name="friendlyName"> The `friendly_name` of the TaskQueue resources to read. </param>
         /// <param name="evaluateWorkerAttributes"> The attributes of the Workers to read. Returns the TaskQueues with Workers that match the attributes specified in this parameter. </param>
         /// <param name="workerSid"> The SID of the Worker with the TaskQueue resources to read. </param>
+        /// <param name="ordering"> Sorting parameter for TaskQueues </param>
         /// <param name="pageSize"> How many resources to return in each list page. The default is 50, and the maximum is 1000. </param>
         /// <param name="client"> Client to make requests to Twilio </param>
         /// <param name="limit"> Record limit </param>
@@ -366,11 +369,12 @@ namespace Twilio.Rest.Taskrouter.V1.Workspace
                                                                                              string friendlyName = null,
                                                                                              string evaluateWorkerAttributes = null,
                                                                                              string workerSid = null,
+                                                                                             string ordering = null,
                                                                                              int? pageSize = null,
                                                                                              long? limit = null,
                                                                                              ITwilioRestClient client = null)
         {
-            var options = new ReadTaskQueueOptions(pathWorkspaceSid){ FriendlyName = friendlyName, EvaluateWorkerAttributes = evaluateWorkerAttributes, WorkerSid = workerSid, PageSize = pageSize, Limit = limit};
+            var options = new ReadTaskQueueOptions(pathWorkspaceSid){ FriendlyName = friendlyName, EvaluateWorkerAttributes = evaluateWorkerAttributes, WorkerSid = workerSid, Ordering = ordering, PageSize = pageSize, Limit = limit};
             return await ReadAsync(options, client);
         }
         #endif