+++ src/Twilio/Rest/Verify/V2/TemplateResource.cs
@@ -188,7 +188,7 @@ namespace Twilio.Rest.Verify.V2
         [JsonProperty("friendly_name")]
         public string FriendlyName { get; private set; }
 
-        ///<summary> The channels </summary> 
+        ///<summary> A list of channels that support the Template </summary> 
         [JsonProperty("channels")]
         public List<string> Channels { get; private set; }